Unify code for listing changed files into shared utility
